Isolate licensekey_handler app (#2)
* Remove obsolete checks

Remove part of check_build and check_lint recipes because they excluded
the only package in the workspace making them redundant. Because they
excluded all packages they also caused CI to fail.

Remove check_tests entirely for the same reason.

Remove check_generated_files because there are no sys-crates with to
generate bindings for.

* Explicitly lock licensekey version

Since licensekey does not yet use semver, the risk of `cargo update`
breaking this app is high. To avoid that we lock the dependency to the
specific commit that we have tested.

* Remove recipe from check_build

It is no longer needed because the prerequisites already ensure that
the app can be built.

* Simplify check_doc and check_lint

Compiling for host is not necessary since the app only works on target.

`--workspace` is not needed because we are no longer using `--exclude`.

* Remove bindings target

Because there are no sys crates left to copy bindings for.

* Format Makefile consistently

Using two new lines before sections to make them more distinct and one
new line before subsections and paragraphs.