Replace ad hoc build system
This is modelled after the corresponding changes in `acap-rs` and the
motivation is largely the same; in summary it makes it easier to share
improvements since we don't need to patch `Makefile`s that may have
diverged, and allows more sophisticated integration with Cargo such as
getting the name of test binaries build or using build scripts to
prepare additional files dynamically.

Enumerate workflows in readme

This gives potential users a preview of what is available without
downloading or installing anything.