Remove Python dependency (#3)
To make the template useful to a wider audience we want to make few
assumptions about the environment in which it will be use. Having
dependencies on programs is one such assumption and the goal is to
depend only on Cargo et. al. and Docker.

Replace `## _` "docstrings" because they are meaningless without
mkhelp, which would expand them to the name of the target when
producing a help text e.g. like "Fix format".