Add summary of make targets to readme
This gives potential users a preview of the DX they should expect and
gives actual users glanceable overview.
The word summarized is chosen to communicate that the `Makefile`
contains more detailed documentation.

Replace "<PACKAGE>" with "app" in Makefile to make it clear that the
project deals only with a single app, so choosing a package to build
would not make sense.