Use hello_world as example
This is the most generic example app from acap-rs.

We use one of the examples from acap-rs to minimize the risk of the
template not building after updates to the build system (will be more
relevant once the build system is a rust binary crate that can be
shared as opposed to an ad hoc `Makefile`).

We choose the most generic example to make the template more
approachable and a minimal starting point that users can add to rather
than remove from.