Update tools from acap-rs
Use a tag instead of a commit ID to make it easier for users to know
and communicate what version they are on. One downside of this is that
the supply chain is more vulnerable since the tags can be moved.

Use the same tag for all programs because the other programs don't yet
have their own tags.