chore: Focus `Makefile` and `README.md` on acap-rs tools
I have now used these tools myself for a bit and they mostly work. When
they fail it is usually not in the parts that are specific to
`cargo-acap-sdk` so even though failures can be difficult for non-
contributors to troubleshoot I don't think it will be more difficult
than before this commit.